+++ src/app/admin/sub-category/sub-category.component.ts
@@ -5,6 +5,7 @@ import { SubCategoryService } from 'src/app/services/sub-category.service';
 import { CategoryService } from 'src/app/services/category.service';
 import { CommonService } from 'src/app/services/common.service';
 import { ValidateSlug } from 'src/app/validations/custom.validators';
+import { TokenService } from 'src/app/services/token.service';
 
 @Component({
   selector: 'app-sub-category',
@@ -35,14 +36,18 @@ export class SubCategoryComponent implements OnInit {
 
   search: any = {};
 
+  sessionRole: String = "";
+
   constructor(
     private fb: FormBuilder,
+    private tokenService: TokenService,
     private subCategoryService: SubCategoryService,
     private commonService: CommonService,
     private responseMessage: NzMessageService
   ) { }
 
   ngOnInit() {
+    this.sessionRole = this.tokenService.getUser().role;
     this.setSearch();
     this.searchData();
     this.setValues();
@@ -117,6 +122,10 @@ export class SubCategoryComponent implements OnInit {
   }
 
   submit() {
+    if (this.sessionRole == "vendor") {
+      return;
+    }
+
     this.submitted = true;
     if (!this.categoryForm.valid) {
       return;
@@ -137,6 +146,10 @@ export class SubCategoryComponent implements OnInit {
   }
 
   delete(id) {
+    if (this.sessionRole == "vendor") {
+      return;
+    }
+    
     this.subCategoryService.delete(id).subscribe(
       data => this.handleResponse(data),
       error => this.handleError(error)
